Provide toga_django templatetags
We might wish to do this to in order to decouple the toga_django
app from its user's apps.

This turns out to be somewhat nontrivial -- the pattern with toga
backends is to import lots of stuff when the module is imported, but
it causes circular imports with Django. The simplest solution seems
to be to move the toga backend initialization logic into another
submodule to clean up the __init__.py file. The drawback of this
approach is that it makes for a strange TOGA_PLATFORM env variable
value:

       os.environ.setdefault("TOGA_PLATFORM", "django.toga_config")

Note that "django.toga_config" does not correspond to a real module,
since toga prepends "toga" to this value.